refactor: modify internal API to better isolate envs
