+++ src/dom/index.ts
@@ -1,0 +1,3 @@
+export * from './create-element'
+export * from './update-elements'
+export * from './utils'
